pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.3.6",
    "django>=5.0.2",
    "django-cors-headers>=4.3.1",
    "django-filter>=23.5",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.7",
    "gunicorn>=21.2.0",
    "psycopg2-binary>=2.9.9",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.0.1",
    "redis>=5.0.1",
    "whitenoise>=6.6.0",
]
